requirements.txt
exif
ExifRead
pandas
Pillow